seperated routes&controller to groceries and list
