+++ app.js
@@ -4,6 +4,7 @@ const morgan = require("morgan");
 const mongoose = require("mongoose");
 require("dotenv").config();
 const grocerieRouter = require("./routes/grocerieRoutes");
+const listRouter = require("./routes/listRoutes");
 const app = express();
 const port = 3000;
 const URI = process.env.DB_URI;
@@ -26,3 +27,4 @@ app.use(express.static("public"));
 app.use(express.urlencoded({ extended: true }));
 app.use(express.json());
 app.use("/", grocerieRouter);
+app.use("/list", listRouter);
