+++ controllers/grocerieController.js
@@ -12,15 +12,6 @@ const findAll = async (req, res) => {
   }
 };
 
-const getList = async (req, res) => {
-  try {
-    const savedList = await SavedList.find().populate("item");
-    res.render("list", { savedList });
-  } catch (err) {
-    res.status(500).json({ err: `getList ${err}` });
-  }
-};
-
 const searchItem = async (req, res) => {
   const name = req.params.name;
   try {
@@ -30,20 +21,6 @@ const searchItem = async (req, res) => {
     res.json(foundItems);
   } catch (err) {
     res.status(500).json({ err: `searchItem ${err}` });
-  }
-};
-
-const searchItemOnList = async (req, res) => {
-  const name = req.params.name;
-  try {
-    const savedList = await SavedList.find().populate({
-      path: "item",
-      match: { name: { $regex: new RegExp(`^${name}`), $options: "i" } },
-    });
-    const html = await res.render("list", { savedList });
-    res.send(html);
-  } catch (err) {
-    res.status(500).json({ err: `searchItemOnList ${err}` });
   }
 };
 
@@ -66,32 +43,6 @@ const addItemToList = async (req, res) => {
   }
 };
 
-const checkedNotOnList = async (req, res) => {
-  try {
-    const groceries = await SavedList.find({ checked: true }).populate("item");
-    const promises = groceries.map((item) => {
-      return Grocerie.updateOne(
-        { name: item.item.name },
-        { $set: { onList: false } }
-      );
-    });
-    await Promise.all(promises);
-    res.json();
-  } catch (err) {
-    res.status(500).json({ err: `checkedNotOnList ${err}` });
-  }
-};
-
-const allNotOnList = async (req, res) => {
-  try {
-    const groceries = await SavedList.find({}).populate("item");
-    const deleted = await Grocerie.updateMany({ groceries }, { onList: false });
-    res.json(deleted);
-  } catch (err) {
-    res.status(500).json({ err: `allNotOnList ${err}` });
-  }
-};
-
 const itemAddedToList = async (req, res) => {
   const _id = req.params.id;
   let result;
@@ -107,57 +58,11 @@ const itemAddedToList = async (req, res) => {
   }
 };
 
-const updateList = async (req, res) => {
-  const _id = req.params.id;
-  try {
-    const update = await SavedList.findOneAndUpdate({ _id }, req.body);
-    console.log("UPDATED");
-    res.json(update);
-  } catch (err) {
-    res.status(500).json({ err: `updateList ${err}` });
-  }
-};
 
-const deleteSavedItem = async (req, res) => {
-  const id = req.params.id;
-  try {
-    const result = await SavedList.deleteOne({ _id: id });
-    res.json(result);
-  } catch (err) {
-    res.status(500).json({ err: `deleteSavedItem ${err}` });
-  }
-};
-
-const deleteChecked = async (req, res) => {
-  try {
-    const result = await SavedList.deleteMany({ checked: true });
-    res.json(result);
-  } catch (err) {
-    res.status(500).json({ err: `deleteChecked ${err}` });
-  }
-};
-
-const deleteAllList = async (req, res) => {
-  try {
-    const result = await SavedList.deleteMany({});
-
-    res.json(result);
-  } catch (err) {
-    res.status(500).json({ err: `deleteAllList ${err}` });
-  }
-};
 
 module.exports = {
   findAll,
-  getList,
   searchItem,
-  searchItemOnList,
   addItemToList,
-  checkedNotOnList,
-  allNotOnList,
   itemAddedToList,
-  updateList,
-  deleteSavedItem,
-  deleteChecked,
-  deleteAllList,
 };
